requirements.txt
selenium
pandas
spacy
reportlab
pdfkit
jinja2
python-dotenv